backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
boto3==1.28.57
